+++ src/stat.h
@@ -3,7 +3,7 @@
 
 #include "types.h"
 
-void stat(stats_t *stats, uchar *node_is, uchar nodesl, const char *const *expr,
+void stat(stats_t *stats, uint *node_is, uint nodesl, const char *const *expr,
           const char *str);
 
 #endif
